Development

See merge request unimax/crud!55

+++ src/app/controller/AccessPointZoneController.ts
@@ -1,5 +1,9 @@
 import { DefaultContext } from 'koa'
+import { logUserEvents } from '../enums/logUserEvents.enum'
+import { AccessPoint } from '../model/entity'
 import { AccessPointZone } from '../model/entity/AccessPointZone'
+import { AntipassBack } from '../model/entity/AntipassBack'
+import { Reader } from '../model/entity/Reader'
 export default class AccessPointZoneController {
     /**
      *
@@ -37,6 +41,23 @@ export default class AccessPointZoneController {
      *                      type: string
      *                  people_limits_max:
      *                      type: string
+     *                  antipass_backs:
+     *                      type: object
+     *                      properties:
+     *                          type:
+     *                              type: string
+     *                              enum: [disable, soft, semi_soft, hard, extra_hard]
+     *                              example: disable
+     *                          enable_timer:
+     *                              type: boolean
+     *                              example: false
+     *                          time:
+     *                              type: number
+     *                              example: 60
+     *                          time_type:
+     *                              type: string
+     *                              enum: [seconds, minutes, hours]
+     *                              example: minutes
      *          responses:
      *              '201':
      *                  description: A accessPointZone object
@@ -51,7 +72,13 @@ export default class AccessPointZoneController {
             const req_data = ctx.request.body
             const user = ctx.user
             req_data.company = user.company ? user.company : null
-            ctx.body = await AccessPointZone.addItem(req_data as AccessPointZone)
+
+            const antipass_back = await AntipassBack.addItem(req_data.antipass_backs as AntipassBack)
+            req_data.antipass_back = antipass_back.id
+            const access_point_zone = await AccessPointZone.addItem(req_data as AccessPointZone) as AccessPointZone
+            access_point_zone.antipass_backs = antipass_back
+
+            ctx.body = access_point_zone
         } catch (error) {
             ctx.status = error.status || 400
             ctx.body = error
@@ -99,6 +126,26 @@ export default class AccessPointZoneController {
      *                      type: string
      *                  people_limits_max:
      *                      type: string
+     *                  antipass_backs:
+     *                      type: object
+     *                      properties:
+     *                          id:
+     *                              type: number
+     *                              example: 1
+     *                          type:
+     *                              type: string
+     *                              enum: [disable, soft, semi_soft, hard, extra_hard]
+     *                              example: disable
+     *                          enable_timer:
+     *                              type: boolean
+     *                              example: false
+     *                          time:
+     *                              type: number
+     *                              example: 60
+     *                          time_type:
+     *                              type: string
+     *                              enum: [seconds, minutes, hours]
+     *                              example: minutes
      *          responses:
      *              '201':
      *                  description: A accessPointZone updated object
@@ -119,8 +166,9 @@ export default class AccessPointZoneController {
                 ctx.body = { message: 'something went wrong' }
             } else {
                 const updated = await AccessPointZone.updateItem(req_data as AccessPointZone)
+                const antipass_back_data = await AntipassBack.updateItem(req_data.antipass_backs as AntipassBack)
                 ctx.oldData = updated.old
-                ctx.body = updated.new
+                ctx.body = { ...updated.new, antipass_backs: antipass_back_data.new }
             }
         } catch (error) {
             ctx.status = error.status || 400
@@ -161,9 +209,22 @@ export default class AccessPointZoneController {
     public static async get (ctx: DefaultContext) {
         try {
             const user = ctx.user
-            const where = { id: +ctx.params.id, company: user.company ? user.company : user.company }
-            const relations = ['access_points']
-            ctx.body = await AccessPointZone.getItem(where, relations)
+            // const where = { id: +ctx.params.id, company: user.company ? user.company : user.company }
+            // const relations = ['access_points']
+            // ctx.body = await AccessPointZone.getItem(where, relations)
+
+            const access_point_zone: any = await AccessPointZone.createQueryBuilder('access_point_zone')
+            .leftJoinAndSelect('access_point_zone.access_points', 'access_point', 'access_point.delete_date is null')
+            .leftJoinAndSelect('access_point.acus', 'acu', 'acu.delete_date is null')
+            .leftJoinAndSelect('access_point_zone.antipass_backs', 'antipass_back')
+            .where(`access_point_zone.id = '${+ctx.params.id}'`)
+            .andWhere(`access_point_zone.company = '${user.company ? user.company : null}'`)
+            .getOne()
+            if (!access_point_zone) {
+                ctx.status = 400
+               return ctx.body = { message: 'Invalid AccessPointZone Id' }
+            }
+            ctx.body = access_point_zone
         } catch (error) {
             ctx.status = error.status || 400
             ctx.body = error
@@ -211,7 +272,35 @@ export default class AccessPointZoneController {
             const user = ctx.user
             const where = { id: req_data.id, company: user.company ? user.company : null }
 
-                ctx.body = await AccessPointZone.destroyItem(where)
+            const childs = await AccessPointZone.findOne({ parent_id: req_data.id })
+            if (childs) {
+                ctx.status = 400
+                ctx.body = { message: 'Can\'t remove group with subzones' }
+            } else {
+                const access_points = await AccessPoint.findOne({ access_point_zone: req_data.id })
+                if (access_points) {
+                    ctx.status = 400
+                    ctx.body = { message: 'Can\'t remove group with access points' }
+                } else {
+                    const readers = await Reader.findOne({
+                        where: [
+                            { leaving_zone: req_data.id },
+                            { came_to_zone: req_data.id }
+                        ]
+                    })
+                    if (readers) {
+                        ctx.status = 400
+                        ctx.body = { message: 'Can\'t remove group with readers' }
+                    }
+                    const access_point_zone = await AccessPointZone.findOneOrFail({ where: where })
+                    ctx.body = await AccessPointZone.destroyItem(where)
+                    ctx.logsData = [{
+                        event: logUserEvents.DELETE,
+                        target: `${AccessPointZone.name}/${access_point_zone.name}`,
+                        value: { name: access_point_zone.name }
+                    }]
+                }
+            }
         } catch (error) {
             ctx.status = error.status || 400
             ctx.body = error
@@ -245,6 +334,7 @@ export default class AccessPointZoneController {
             const req_data = ctx.query
             const user = ctx.user
             req_data.where = { company: { '=': user.company ? user.company : null } }
+            req_data.relations = ['antipass_backs']
             ctx.body = await AccessPointZone.getAllItems(req_data)
         } catch (error) {
             ctx.status = error.status || 400
