+++ src/app/controller/AccountGroupController.ts
@@ -1,4 +1,6 @@
 import { DefaultContext } from 'koa'
+import { logUserEvents } from '../enums/logUserEvents.enum'
+import { AccessPointGroup, Admin } from '../model/entity'
 import { AccountGroup } from '../model/entity/AccountGroup'
 export default class AccountGroupController {
     /**
@@ -212,7 +214,28 @@ export default class AccountGroupController {
             const req_data = ctx.request.body
             const user = ctx.user
             const where = { id: req_data.id, company: user.company ? user.company : null }
-            ctx.body = await AccountGroup.destroyItem(where)
+
+            const childs = await AccountGroup.find({ parent_id: req_data.id })
+            if (childs.length) {
+                ctx.status = 400
+                ctx.body = { message: 'Can\'t remove group with childs' }
+            } else {
+                const admins = await Admin.find({ account_group: req_data.id })
+                if (admins.length) {
+                    for (const admin of admins) {
+                        admin.account_group = null
+                        await admin.save()
+                    }
+                } else {
+                    const account_group = await AccountGroup.findOneOrFail({ where: where })
+                    ctx.body = await AccountGroup.destroyItem(where)
+                    ctx.logsData = [{
+                        event: logUserEvents.DELETE,
+                        target: `${AccessPointGroup.name}/${account_group.name}`,
+                        value: { name: account_group.name }
+                    }]
+                }
+            }
         } catch (error) {
             ctx.status = error.status || 400
             ctx.body = error
